+++ pkg/providers/property/data_akamai_cp_code_test.go
@@ -193,6 +193,9 @@ func TestDSCPCode(t *testing.T) {
 	t.Run("group not found in state", func(t *testing.T) {
 		t.Parallel()
 		client := edgegrid.NewTestClient()
+
+		expectGetProducts(client.PAPI, "ctr_11", []string{"prd_1", "prd_2", "prd_3"}).Once()
+
 		client.PAPI.On("GetCPCodes",
 			testutils.MockContext, papi.GetCPCodesRequest{ContractID: "ctr_11", GroupID: "grp_22"},
 		).Return(&papi.GetCPCodesResponse{CPCodes: papi.CPCodeItems{Items: []papi.CPCode{{