+++ pkg/providers/property/testdata/TestResCPCode/existing_cp_code.tf
@@ -1,0 +1,14 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_cp_code" "test" {
+  name        = "test cpcode"
+  contract_id = "ctr_1"
+  group_id    = "grp_1"
+  product_id  = "prd_1"
+  # this is to avoid recreating the resource when product_id changes
+  lifecycle {
+    ignore_changes = [product_id]
+  }
+}